requirements.txt
absl-py==2.3.1
accelerate==1.10.1
annotated-types==0.7.0
anyio==4.10.0
astunparse==1.6.3
bitsandbytes==0.47.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
colorama==0.4.6
datauri==1.0.0
einops==0.8.1
fastapi==0.116.1
filelock==3.19.1
flatbuffers==25.2.10
fsspec==2025.9.0
gast==0.6.0
google-pasta==0.2.0
grpcio==1.74.0
h11==0.16.0
h5py==3.14.0
hf_transfer==0.1.9
huggingface-hub==0.34.4
idna==3.10
Jinja2==3.1.6
keras==3.11.3
libclang==18.1.1
loguru==0.7.3
Markdown==3.9
markdown-it-py==4.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
ml_dtypes==0.5.3
mpmath==1.3.0
namex==0.1.0
networkx==3.5
numpy==2.2.6
opencv-python==4.12.0.88
opt_einsum==3.4.0
optree==0.17.0
packaging==25.0
pillow==11.3.0
protobuf==6.32.0
psutil==7.0.0
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
PyYAML==6.0.2
regex==2025.9.1
requests==2.32.5
rich==14.1.0
safetensors==0.6.2
six==1.17.0
sniffio==1.3.1
sse-starlette==3.0.2
starlette==0.47.3
sympy==1.13.1
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensorflow==2.20.0
termcolor==3.1.0
tokenizers==0.22.0
torch==2.6.0
torchvision==0.21.0
tqdm==4.67.1
transformers==4.56.1
typing-inspection==0.4.1
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.35.0
Werkzeug==3.1.3
win32_setctime==1.2.0
wrapt==1.17.3
